refactor: Add shared categorical chart hook
