+++ packages/api/src/mcp/__tests__/dashboards/queryTile.test.ts
@@ -1,4 +1,12 @@
+import { SourceKind } from '@hyperdx/common-utils/dist/types';
+
+import {
+  bulkInsertLogs,
+  DEFAULT_DATABASE,
+  DEFAULT_LOGS_TABLE,
+} from '@/fixtures';
 import { callTool, getFirstText } from '@/mcp/__tests__/mcpTestUtils';
+import { Source } from '@/models/source';
 
 import { setupDashboardTests } from './setup';
 
@@ -190,6 +198,108 @@ describe('MCP Dashboard Tools - clickstack_query_tile', () => {
       expect(result.isError).toBeFalsy();
       const parsed = JSON.parse(getFirstText(result));
       expect(parsed.warnings).toBeUndefined();
+    });
+  });
+
+  describe('categorical (bar) tile series limit', () => {
+    const now = new Date();
+    const tenMinAgo = new Date(now.getTime() - 10 * 60 * 1000);
+
+    // Five distinct services with descending row counts, so a limit keeps a
+    // strict, deterministic top-N subset.
+    const SERVICE_COUNTS: [string, number][] = [
+      ['svc-a', 5],
+      ['svc-b', 4],
+      ['svc-c', 3],
+      ['svc-d', 2],
+      ['svc-e', 1],
+    ];
+
+    let logSourceId: string;
+
+    beforeEach(async () => {
+      const logSource = await Source.create({
+        kind: SourceKind.Log,
+        team: ctx.team._id,
+        from: {
+          databaseName: DEFAULT_DATABASE,
+          tableName: DEFAULT_LOGS_TABLE,
+        },
+        timestampValueExpression: 'Timestamp',
+        connection: ctx.connection._id,
+        name: 'Bar Limit Logs',
+        bodyExpression: 'Body',
+        severityTextExpression: 'SeverityText',
+      });
+      logSourceId = logSource._id.toString();
+
+      const logs: Parameters<typeof bulkInsertLogs>[0] = [];
+      SERVICE_COUNTS.forEach(([serviceName, count], svcIdx) => {
+        for (let i = 0; i < count; i++) {
+          logs.push({
+            Body: `bar limit log ${serviceName} ${i}`,
+            ServiceName: serviceName,
+            SeverityText: 'INFO',
+            Timestamp: new Date(tenMinAgo.getTime() + (svcIdx * 100 + i) * 10),
+          });
+        }
+      });
+      await bulkInsertLogs(logs);
+    });
+
+    const saveBarDashboard = async (limit?: number) => {
+      const createResult = await callTool(
+        ctx.client!,
+        'clickstack_save_dashboard',
+        {
+          name: `Bar Limit Dashboard ${limit ?? 'none'}`,
+          tiles: [
+            {
+              name: 'Bars by service',
+              config: {
+                displayType: 'bar',
+                sourceId: logSourceId,
+                select: [{ aggFn: 'count' }],
+                groupBy: 'ServiceName',
+                ...(limit != null ? { limit } : {}),
+              },
+            },
+          ],
+        },
+      );
+      expect(createResult.isError).toBeFalsy();
+      return JSON.parse(getFirstText(createResult));
+    };
+
+    const queryTileRows = async (dashboard: any) => {
+      const result = await callTool(ctx.client!, 'clickstack_query_tile', {
+        dashboardId: dashboard.id,
+        tileId: dashboard.tiles[0].id,
+        startTime: new Date(now.getTime() - 15 * 60 * 1000).toISOString(),
+        endTime: new Date(now.getTime() + 60 * 1000).toISOString(),
+      });
+      expect(result.isError).toBeFalsy();
+      const parsed = JSON.parse(getFirstText(result));
+      return parsed.result.data as Record<string, string>[];
+    };
+
+    it('returns every group when no limit is set', async () => {
+      const dashboard = await saveBarDashboard();
+      const rows = await queryTileRows(dashboard);
+      expect(rows).toHaveLength(SERVICE_COUNTS.length);
+    });
+
+    it('caps the bars to the series limit, keeping the largest groups', async () => {
+      const dashboard = await saveBarDashboard(3);
+      const rows = await queryTileRows(dashboard);
+
+      // The limit must actually reduce the result below the full set.
+      expect(rows).toHaveLength(3);
+      expect(rows.length).toBeLessThan(SERVICE_COUNTS.length);
+
+      // And it must keep the top-3 by count, not an arbitrary subset.
+      const services = rows.map(r => r.ServiceName).sort();
+      expect(services).toEqual(['svc-a', 'svc-b', 'svc-c']);
     });
   });
 });
