+++ packages/app/src/ChartUtils.tsx
@@ -14,6 +14,7 @@ import { isMetricChartConfig } from '@hyperdx/common-utils/dist/core/renderChart
 import {
   convertDateRangeToGranularityString,
   convertGranularityToSeconds,
+  convertToCategoricalChartConfig,
   getAlignedDateRange,
   Granularity,
 } from '@hyperdx/common-utils/dist/core/utils';
@@ -1138,60 +1139,8 @@ export function convertToNumberChartConfig(
   return omit(config, ['granularity', 'groupBy']);
 }
 
-export function convertToCategoricalChartConfig(
-  config: BuilderChartConfigWithOptTimestamp,
-): BuilderChartConfigWithOptTimestamp {
-  const convertedConfig = structuredClone(omit(config, ['granularity']));
-
-  // Pie/bar charts interpret `seriesLimit` as a plain SQL LIMIT on the
-  // number of slices/bars (the __hdx_series_limit ranking CTE it drives on
-  // time charts is gated on granularity, which is not used for categorical charts).
-  if (
-    convertedConfig.seriesLimit != null &&
-    convertedConfig.limit?.limit == null
-  ) {
-    convertedConfig.limit = { limit: convertedConfig.seriesLimit };
-    delete convertedConfig.seriesLimit;
-  }
-
-  // When a series limit is set, order by the first aggregated value
-  // descending (with the group as a stable tiebreak) so the limit
-  // deterministically keeps the largest slices/bars.
-  //
-  // The value column is
-  // referenced by its alias — the renderer emits `expr AS "alias"` — so
-  // inject the metrics-convention 'Value' alias when the user didn't set
-  // one. Without a limit, ordering is irrelevant: the chart sorts
-  // client-side by value.
-  if (
-    convertedConfig.limit?.limit != null &&
-    convertedConfig.orderBy == null &&
-    Array.isArray(convertedConfig.select) &&
-    convertedConfig.select.length > 0 &&
-    typeof convertedConfig.groupBy === 'string' &&
-    convertedConfig.seriesReturnType !== 'ratio'
-  ) {
-    const firstSelect = convertedConfig.select[0];
-    if (!firstSelect.alias?.trim()) {
-      firstSelect.alias = 'Value';
-    }
-    // Quote the alias as a ClickHouse identifier, doubling any embedded
-    // double quotes so aliases like `Request "Count"` are escaped correctly.
-    const quotedAlias = `"${firstSelect.alias.trim().replace(/"/g, '""')}"`;
-    convertedConfig.orderBy = [
-      {
-        valueExpression: quotedAlias,
-        ordering: 'DESC',
-      },
-      {
-        valueExpression: convertedConfig.groupBy,
-        ordering: 'ASC' as const,
-      },
-    ];
-  }
-
-  return convertedConfig;
-}
+// The pie/bar seriesLimit→LIMIT translation lives in common-utils
+export { convertToCategoricalChartConfig };
 
 export function convertToTableChartConfig(
   config: BuilderChartConfigWithOptTimestamp,
