+++ packages/app/src/__tests__/ChartUtils.test.ts
@@ -5,7 +5,6 @@ import {
 } from '@hyperdx/common-utils/dist/types';
 
 import {
-  convertToCategoricalChartConfig,
   convertToNumberChartConfig,
   convertToTableChartConfig,
   convertToTimeChartConfig,
@@ -856,136 +855,6 @@ describe('ChartUtils', () => {
 
       expect(convertedConfig.granularity).toBeUndefined();
       expect(convertedConfig.groupBy).toBeUndefined();
-    });
-  });
-
-  describe('convertToCategoricalChartConfig', () => {
-    const dateRange: [Date, Date] = [
-      new Date('2025-11-26T00:00:00Z'),
-      new Date('2025-11-27T00:00:00Z'),
-    ];
-
-    it('should remove granularity but keep groupBy', () => {
-      const config = {
-        granularity: '5 minute',
-        groupBy: 'ServiceName',
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.granularity).toBeUndefined();
-      expect(convertedConfig.groupBy).toEqual('ServiceName');
-    });
-
-    it('should not inject a limit or orderBy when no seriesLimit is set', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        groupBy: 'ServiceName',
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.orderBy).toBeUndefined();
-      expect(convertedConfig.limit).toBeUndefined();
-    });
-
-    it('should apply seriesLimit as a limit ordered by the first value descending', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        groupBy: 'ServiceName',
-        seriesLimit: 5,
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.select[0]).toMatchObject({ alias: 'Value' });
-      expect(convertedConfig.orderBy).toEqual([
-        { valueExpression: '"Value"', ordering: 'DESC' },
-        { valueExpression: 'ServiceName', ordering: 'ASC' },
-      ]);
-      expect(convertedConfig.limit).toEqual({ limit: 5 });
-    });
-
-    it('should order by the existing alias when the first series has one', () => {
-      const config = {
-        select: [
-          { aggFn: 'count', valueExpression: '', alias: 'Request "Count"' },
-        ],
-        groupBy: 'ServiceName',
-        seriesLimit: 3,
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.select[0]).toMatchObject({
-        alias: 'Request "Count"',
-      });
-      expect(convertedConfig.orderBy).toEqual([
-        { valueExpression: '"Request ""Count"""', ordering: 'DESC' },
-        { valueExpression: 'ServiceName', ordering: 'ASC' },
-      ]);
-      expect(convertedConfig.limit).toEqual({ limit: 3 });
-    });
-
-    it('should preserve an explicit orderBy over the value-descending default', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        groupBy: 'ServiceName',
-        seriesLimit: 5,
-        orderBy: 'ServiceName ASC',
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.orderBy).toEqual('ServiceName ASC');
-      expect(convertedConfig.limit).toEqual({ limit: 5 });
-      expect(convertedConfig.select[0]).not.toHaveProperty('alias');
-    });
-
-    it('should preserve an explicit limit over seriesLimit', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        groupBy: 'ServiceName',
-        seriesLimit: 5,
-        limit: { limit: 2 },
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.limit).toEqual({ limit: 2 });
-    });
-
-    it('should not inject an orderBy when there is no groupBy', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        seriesLimit: 5,
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      const convertedConfig = convertToCategoricalChartConfig(config);
-
-      expect(convertedConfig.orderBy).toBeUndefined();
-    });
-
-    it('should not mutate the input config', () => {
-      const config = {
-        select: [{ aggFn: 'count', valueExpression: '' }],
-        groupBy: 'ServiceName',
-        seriesLimit: 5,
-        dateRange,
-      } as BuilderChartConfigWithDateRange;
-
-      convertToCategoricalChartConfig(config);
-
-      expect(config.select[0]).not.toHaveProperty('alias');
-      expect(config.orderBy).toBeUndefined();
-      expect(config.limit).toBeUndefined();
     });
   });
