+++ packages/app/src/components/DBPieChart.tsx
@@ -1,26 +1,17 @@
-import { memo, useMemo } from 'react';
+import { memo } from 'react';
 import { Cell, Pie, PieChart, ResponsiveContainer, Tooltip } from 'recharts';
-import { isBuilderChartConfig } from '@hyperdx/common-utils/dist/guards';
-import { ChartConfigWithOptTimestamp } from '@hyperdx/common-utils/dist/types';
 import { Box, Flex, ScrollArea, Text } from '@mantine/core';
 
-import {
-  buildMVDateRangeIndicator,
-  convertToCategoricalChartConfig,
-  formatResponseForCategoricalChart,
-} from '@/ChartUtils';
-import { useQueriedChartConfig } from '@/hooks/useChartConfig';
-import { useMVOptimizationExplanation } from '@/hooks/useMVOptimizationExplanation';
-import { useSingleSeriesNumberFormat, useSource } from '@/source';
 import type { NumberFormat } from '@/types';
-import { formatNumber, getColorProps, truncateMiddle } from '@/utils';
+import { formatNumber, truncateMiddle } from '@/utils';
 
+import {
+  CategoricalChartProps,
+  useCategoricalChart,
+} from './charts/CategoricalChart';
 import ChartContainer from './charts/ChartContainer';
-import ChartErrorState, {
-  ChartErrorStateVariant,
-} from './charts/ChartErrorState';
+import ChartErrorState from './charts/ChartErrorState';
 import { ChartTooltipContainer, ChartTooltipItem } from './charts/ChartTooltip';
-import MVOptimizationIndicator from './MaterializedViews/MVOptimizationIndicator';
 
 const PieChartTooltip = memo(
   ({
@@ -96,105 +87,21 @@ const PieChartLegend = memo(
   },
 );
 
-export const DBPieChart = ({
-  config,
-  title,
-  enabled = true,
-  queryKeyPrefix,
-  showMVOptimizationIndicator = true,
-  toolbarPrefix,
-  toolbarSuffix,
-  errorVariant,
-}: {
-  config: ChartConfigWithOptTimestamp;
-  title?: React.ReactNode;
-  enabled?: boolean;
-  queryKeyPrefix?: string;
-  showMVOptimizationIndicator?: boolean;
-  toolbarPrefix?: React.ReactNode[];
-  toolbarSuffix?: React.ReactNode[];
-  errorVariant?: ChartErrorStateVariant;
-}) => {
-  const { data: source } = useSource({
-    id: config.source,
-  });
-
-  const queriedConfig = useMemo(() => {
-    return isBuilderChartConfig(config)
-      ? convertToCategoricalChartConfig(config)
-      : config;
-  }, [config]);
-
-  const resolvedNumberFormat = useSingleSeriesNumberFormat(queriedConfig);
-
-  const builderQueriedConfig = isBuilderChartConfig(queriedConfig)
-    ? queriedConfig
-    : undefined;
-  const { data: mvOptimizationData } =
-    useMVOptimizationExplanation(builderQueriedConfig);
-
-  const { data, isLoading, isError, error } = useQueriedChartConfig(
-    queriedConfig,
-    {
-      placeholderData: (prev: any) => prev,
-      queryKey: [queryKeyPrefix, queriedConfig],
-      enabled,
-    },
-  );
-
-  const toolbarItemsMemo = useMemo(() => {
-    const allToolbarItems = [];
-
-    if (toolbarPrefix && toolbarPrefix.length > 0) {
-      allToolbarItems.push(...toolbarPrefix);
-    }
-
-    if (source && showMVOptimizationIndicator && builderQueriedConfig) {
-      allToolbarItems.push(
-        <MVOptimizationIndicator
-          key="db-table-chart-mv-indicator"
-          config={builderQueriedConfig}
-          source={source}
-          variant="icon"
-        />,
-      );
-    }
-
-    const dateRangeIndicator = buildMVDateRangeIndicator({
-      mvOptimizationData,
-      originalDateRange: queriedConfig.dateRange,
-    });
-
-    if (dateRangeIndicator) {
-      allToolbarItems.push(dateRangeIndicator);
-    }
-
-    if (toolbarSuffix && toolbarSuffix.length > 0) {
-      allToolbarItems.push(...toolbarSuffix);
-    }
-
-    return allToolbarItems;
-  }, [
-    toolbarPrefix,
-    toolbarSuffix,
-    source,
-    showMVOptimizationIndicator,
-    mvOptimizationData,
-    queriedConfig,
-    builderQueriedConfig,
-  ]);
-
-  const [pieChartData, responseFormatError] = useMemo(() => {
-    if (!data) return [[], null];
-    try {
-      return [formatResponseForCategoricalChart(data, getColorProps), null];
-    } catch (error) {
-      return [[], error instanceof Error ? error : new Error(String(error))];
-    }
-  }, [data]);
+export const DBPieChart = (props: CategoricalChartProps) => {
+  const { title, errorVariant } = props;
+  const {
+    resolvedNumberFormat,
+    toolbarItems,
+    data,
+    isLoading,
+    isError,
+    error,
+    chartData,
+    responseFormatError,
+  } = useCategoricalChart(props);
 
   return (
-    <ChartContainer title={title} toolbarItems={toolbarItemsMemo}>
+    <ChartContainer title={title} toolbarItems={toolbarItems}>
       {isLoading && !data ? (
         <div className="d-flex h-100 w-100 align-items-center justify-content-center text-muted">
           Loading Chart Data...
@@ -224,12 +131,12 @@ export const DBPieChart = ({
               <Pie
                 cx="50%"
                 cy="50%"
-                data={pieChartData}
+                data={chartData}
                 dataKey="value"
                 fill="#8884d8"
                 nameKey="label"
               >
-                {pieChartData.map(entry => (
+                {chartData.map(entry => (
                   <Cell key={entry.label} fill={entry.color} stroke="none" />
                 ))}
               </Pie>
@@ -241,7 +148,7 @@ export const DBPieChart = ({
             </PieChart>
           </ResponsiveContainer>
           <PieChartLegend
-            data={pieChartData}
+            data={chartData}
             numberFormat={resolvedNumberFormat}
           />
         </Flex>
